requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
dnspython==2.7.0
fastapi==0.115.14
graphql-core==3.2.6
h11==0.16.0
idna==3.10
libcst==1.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
motor==3.7.1
packaging==25.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
strawberry-graphql==0.275.5
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
websockets==15.0.1
